Revert "Add FindCommand"
This reverts commit 62a06ed0788320bc5dde39c7ff30ee9e8f362ace.
