+++ src/main/java/gigi/Gigi.java
@@ -1,10 +1,11 @@
-package java;
+package main.java.gigi;
 
-import java.commands.Command;
-import java.exceptions.GigiException;
-import java.storage.Storage;
-import java.tasks.Tasklist;
-import java.ui.Ui;
+import gigi.commands.Command;
+import gigi.exceptions.GigiException;
+import gigi.main.Parser;
+import gigi.storage.Storage;
+import gigi.tasks.Tasklist;
+import gigi.ui.Ui;
 
 public class Gigi {
     private static final String FILE_PATH = "./data/Gigi.txt";
