+++ src/main/java/gigi/META-INF/MANIFEST.MF
