+++ src/main/java/gigi/exceptions/GigiException.java
@@ -1,4 +1,4 @@
-package java.exceptions;
+package gigi.exceptions;
 
 public class GigiException extends Exception{
