+++ src/main/java/gigi/tasks/DeadlinesTest.java
@@ -1,8 +1,9 @@
-package java.tasks;
+package gigi.tasks;
 
 import org.junit.jupiter.api.Test;
 
 import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
 
 import static org.junit.jupiter.api.Assertions.*;
