+++ src/main/java/gigi/tasks/Events.java
@@ -1,4 +1,4 @@
-package java.tasks;
+package gigi.tasks;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
