+++ src/main/java/gigi/tasks/ToDosTest.java
@@ -1,4 +1,4 @@
-package java.tasks;
+package gigi.tasks;
 
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.*;
